mandatory and subject

+++ pipex/ft_strncmp.c
@@ -6,13 +6,13 @@
 /*   By: sbart <sbart@student.21-school.ru>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/13 18:32:56 by sbart             #+#    #+#             */
-/*   Updated: 2022/03/13 18:33:19 by sbart            ###   ########.fr       */
+/*   Updated: 2022/03/15 17:50:52 by sbart            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "pipex.h"
 
-int	ft_strncmp(const char *s1, const char *s2, size_t n)
+int	ft_strncmp(const char *s1, const char *s2, int n)
 {
 	while (n > 0 && *s1 && *s2)
 	{
